+++ src/torchmetrics/classification/accuracy.py
@@ -18,6 +18,7 @@ from typing_extensions import Literal
 
 from torchmetrics.functional.classification.accuracy import _accuracy_reduce
 from torchmetrics.metric import Metric
+from torchmetrics.utilities.enums import ClassificationTask
 from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE
 from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE, plot_single_or_multi_val
 
@@ -149,7 +150,7 @@ class BinaryAccuracy(BinaryStatScores):
 
 
 class MulticlassAccuracy(MulticlassStatScores):
-    r"""Compute `Accuracy`_ for multiclass tasks:
+    r"""Compute `Accuracy`_ for multiclass tasks.
 
     .. math::
         \text{Accuracy} = \frac{1}{N}\sum_i^N 1(y_i = \hat{y}_i)
@@ -298,7 +299,7 @@ class MulticlassAccuracy(MulticlassStatScores):
 
 
 class MultilabelAccuracy(MultilabelStatScores):
-    r"""Compute `Accuracy`_ for multilabel tasks:
+    r"""Compute `Accuracy`_ for multilabel tasks.
 
     .. math::
         \text{Accuracy} = \frac{1}{N}\sum_i^N 1(y_i = \hat{y}_i)
@@ -490,18 +491,17 @@ class Accuracy:
         validate_args: bool = True,
         **kwargs: Any,
     ) -> Metric:
+        """Initialize task metric."""
+        task = ClassificationTask.from_str(task)
         kwargs.update(
             {"multidim_average": multidim_average, "ignore_index": ignore_index, "validate_args": validate_args}
         )
-        if task == "binary":
+        if task == ClassificationTask.BINARY:
             return BinaryAccuracy(threshold, **kwargs)
-        if task == "multiclass":
+        if task == ClassificationTask.MULTICLASS:
             assert isinstance(num_classes, int)
             assert isinstance(top_k, int)
             return MulticlassAccuracy(num_classes, top_k, average, **kwargs)
-        if task == "multilabel":
+        if task == ClassificationTask.MULTILABEL:
             assert isinstance(num_labels, int)
             return MultilabelAccuracy(num_labels, threshold, average, **kwargs)
-        raise ValueError(
-            f"Expected argument `task` to either be `'binary'`, `'multiclass'` or `'multilabel'` but got {task}"
-        )