+++ src/torchmetrics/classification/matthews_corrcoef.py
@@ -19,6 +19,7 @@ from typing_extensions import Literal
 from torchmetrics.classification import BinaryConfusionMatrix, MulticlassConfusionMatrix, MultilabelConfusionMatrix
 from torchmetrics.functional.classification.matthews_corrcoef import _matthews_corrcoef_reduce
 from torchmetrics.metric import Metric
+from torchmetrics.utilities.enums import ClassificationTask
 
 
 class BinaryMatthewsCorrCoef(BinaryConfusionMatrix):
@@ -79,6 +80,7 @@ class BinaryMatthewsCorrCoef(BinaryConfusionMatrix):
         super().__init__(threshold, ignore_index, normalize=None, validate_args=validate_args, **kwargs)
 
     def compute(self) -> Tensor:
+        """Compute metric."""
         return _matthews_corrcoef_reduce(self.confmat)
 
 
@@ -143,6 +145,7 @@ class MulticlassMatthewsCorrCoef(MulticlassConfusionMatrix):
         super().__init__(num_classes, ignore_index, normalize=None, validate_args=validate_args, **kwargs)
 
     def compute(self) -> Tensor:
+        """Compute metric."""
         return _matthews_corrcoef_reduce(self.confmat)
 
 
@@ -206,6 +209,7 @@ class MultilabelMatthewsCorrCoef(MultilabelConfusionMatrix):
         super().__init__(num_labels, threshold, ignore_index, normalize=None, validate_args=validate_args, **kwargs)
 
     def compute(self) -> Tensor:
+        """Compute metric."""
         return _matthews_corrcoef_reduce(self.confmat)
 
 
@@ -237,15 +241,14 @@ class MatthewsCorrCoef:
         validate_args: bool = True,
         **kwargs: Any,
     ) -> Metric:
+        """Initialize task metric."""
+        task = ClassificationTask.from_str(task)
         kwargs.update({"ignore_index": ignore_index, "validate_args": validate_args})
-        if task == "binary":
+        if task == ClassificationTask.BINARY:
             return BinaryMatthewsCorrCoef(threshold, **kwargs)
-        if task == "multiclass":
+        if task == ClassificationTask.MULTICLASS:
             assert isinstance(num_classes, int)
             return MulticlassMatthewsCorrCoef(num_classes, **kwargs)
-        if task == "multilabel":
+        if task == ClassificationTask.MULTILABEL:
             assert isinstance(num_labels, int)
             return MultilabelMatthewsCorrCoef(num_labels, threshold, **kwargs)
-        raise ValueError(
-            f"Expected argument `task` to either be `'binary'`, `'multiclass'` or `'multilabel'` but got {task}"
-        )