+++ src/torchmetrics/classification/roc.py
@@ -36,6 +36,7 @@ from torchmetrics.functional.classification.roc import (
 )
 from torchmetrics.metric import Metric
 from torchmetrics.utilities.data import dim_zero_cat
+from torchmetrics.utilities.enums import ClassificationTask
 from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE
 from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE, plot_binary_roc_curve
 
@@ -112,6 +113,7 @@ class BinaryROC(BinaryPrecisionRecallCurve):
     full_state_update: bool = False
 
     def compute(self) -> Tuple[Tensor, Tensor, Tensor]:
+        """Compute metric."""
         state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)] if self.thresholds is None else self.confmat
         return _binary_roc_compute(state, self.thresholds)
 
@@ -260,6 +262,7 @@ class MulticlassROC(MulticlassPrecisionRecallCurve):
     full_state_update: bool = False
 
     def compute(self) -> Tuple[Tensor, Tensor, Tensor]:
+        """Compute metric."""
         state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)] if self.thresholds is None else self.confmat
         return _multiclass_roc_compute(state, self.num_classes, self.thresholds)
 
@@ -363,6 +366,7 @@ class MultilabelROC(MultilabelPrecisionRecallCurve):
     full_state_update: bool = False
 
     def compute(self) -> Tuple[Tensor, Tensor, Tensor]:
+        """Compute metric."""
         state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)] if self.thresholds is None else self.confmat
         return _multilabel_roc_compute(state, self.num_labels, self.thresholds, self.ignore_index)
 
@@ -438,15 +442,14 @@ class ROC:
         validate_args: bool = True,
         **kwargs: Any,
     ) -> Metric:
+        """Initialize task metric."""
+        task = ClassificationTask.from_str(task)
         kwargs.update({"thresholds": thresholds, "ignore_index": ignore_index, "validate_args": validate_args})
-        if task == "binary":
+        if task == ClassificationTask.BINARY:
             return BinaryROC(**kwargs)
-        if task == "multiclass":
+        if task == ClassificationTask.MULTICLASS:
             assert isinstance(num_classes, int)
             return MulticlassROC(num_classes, **kwargs)
-        if task == "multilabel":
+        if task == ClassificationTask.MULTILABEL:
             assert isinstance(num_labels, int)
             return MultilabelROC(num_labels, **kwargs)
-        raise ValueError(
-            f"Expected argument `task` to either be `'binary'`, `'multiclass'` or `'multilabel'` but got {task}"
-        )