+++ src/torchmetrics/classification/specificity_at_sensitivity.py
@@ -30,7 +30,8 @@ from torchmetrics.functional.classification.specificity_at_sensitivity import (
     _multilabel_specificity_at_sensitivity_compute,
 )
 from torchmetrics.metric import Metric
-from torchmetrics.utilities.data import dim_zero_cat
+from torchmetrics.utilities.data import dim_zero_cat as _cat
+from torchmetrics.utilities.enums import ClassificationTask
 
 
 class BinarySpecificityAtSensitivity(BinaryPrecisionRecallCurve):
@@ -108,10 +109,8 @@ class BinarySpecificityAtSensitivity(BinaryPrecisionRecallCurve):
         self.min_sensitivity = min_sensitivity
 
     def compute(self) -> Tuple[Tensor, Tensor]:  # type: ignore[override]
-        if self.thresholds is None:
-            state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)]  # type: ignore
-        else:
-            state = self.confmat
+        """Compute metric."""
+        state = [_cat(self.preds), _cat(self.target)] if self.thresholds is None else self.confmat  # type: ignore
         return _binary_specificity_at_sensitivity_compute(state, self.thresholds, self.min_sensitivity)  # type: ignore
 
 
@@ -201,10 +200,8 @@ class MulticlassSpecificityAtSensitivity(MulticlassPrecisionRecallCurve):
         self.min_sensitivity = min_sensitivity
 
     def compute(self) -> Tuple[Tensor, Tensor]:  # type: ignore
-        if self.thresholds is None:
-            state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)]  # type: ignore
-        else:
-            state = self.confmat
+        """Compute metric."""
+        state = [_cat(self.preds), _cat(self.target)] if self.thresholds is None else self.confmat  # type: ignore
         return _multiclass_specificity_at_sensitivity_compute(
             state, self.num_classes, self.thresholds, self.min_sensitivity  # type: ignore
         )
@@ -296,10 +293,8 @@ class MultilabelSpecificityAtSensitivity(MultilabelPrecisionRecallCurve):
         self.min_sensitivity = min_sensitivity
 
     def compute(self) -> Tuple[Tensor, Tensor]:  # type: ignore[override]
-        if self.thresholds is None:
-            state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)]  # type: ignore
-        else:
-            state = self.confmat
+        """Compute metric."""
+        state = [_cat(self.preds), _cat(self.target)] if self.thresholds is None else self.confmat  # type: ignore
         return _multilabel_specificity_at_sensitivity_compute(
             state, self.num_labels, self.thresholds, self.ignore_index, self.min_sensitivity  # type: ignore
         )
@@ -327,18 +322,17 @@ class SpecificityAtSensitivity:
         validate_args: bool = True,
         **kwargs: Any,
     ) -> Metric:
-        if task == "binary":
+        """Initialize task metric."""
+        task = ClassificationTask.from_str(task)
+        if task == ClassificationTask.BINARY:
             return BinarySpecificityAtSensitivity(min_sensitivity, thresholds, ignore_index, validate_args, **kwargs)
-        if task == "multiclass":
+        if task == ClassificationTask.MULTICLASS:
             assert isinstance(num_classes, int)
             return MulticlassSpecificityAtSensitivity(
                 num_classes, min_sensitivity, thresholds, ignore_index, validate_args, **kwargs
             )
-        if task == "multilabel":
+        if task == ClassificationTask.MULTILABEL:
             assert isinstance(num_labels, int)
             return MultilabelSpecificityAtSensitivity(
                 num_labels, min_sensitivity, thresholds, ignore_index, validate_args, **kwargs
             )
-        raise ValueError(
-            f"Expected argument `task` to either be `'binary'`, `'multiclass'` or `'multilabel'` but got {task}"
-        )