+++ src/torchmetrics/image/lpip.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import os
 from typing import Any, List
 
 import torch
@@ -19,10 +20,17 @@ from torch.nn import Module
 from typing_extensions import Literal
 
 from torchmetrics.metric import Metric
+from torchmetrics.utilities.checks import _SKIP_SLOW_DOCTEST, _try_proceed_with_timeout
 from torchmetrics.utilities.imports import _LPIPS_AVAILABLE
 
 if _LPIPS_AVAILABLE:
     from lpips import LPIPS as _LPIPS
+
+    def _download_lpips() -> None:
+        _LPIPS(pretrained=True, net="vgg")
+
+    if _SKIP_SLOW_DOCTEST and not _try_proceed_with_timeout(_download_lpips):
+        __doctest_skip__ = ["LearnedPerceptualImagePatchSimilarity", "LPIPS"]
 else:
 
     class _LPIPS(Module):  # type: ignore
@@ -32,6 +40,8 @@ else:
 
 
 class NoTrainLpips(_LPIPS):
+    """Wrapper to make sure LPIPS never leaves evaluation mode."""
+
     def train(self, mode: bool) -> "NoTrainLpips":
         """the network should not be able to be switched away from evaluation mode."""
         return super().train(False)