+++ tests/unittests/utilities/test_plot.py
@@ -19,6 +19,15 @@ import numpy as np
 import pytest
 import torch
 
+from torchmetrics.functional import (
+    scale_invariant_signal_distortion_ratio,
+    scale_invariant_signal_noise_ratio,
+    signal_distortion_ratio,
+    signal_noise_ratio,
+)
+from torchmetrics.functional.audio import short_time_objective_intelligibility
+from torchmetrics.functional.audio.pesq import perceptual_evaluation_speech_quality
+from torchmetrics.functional.audio.pit import permutation_invariant_training
 from torchmetrics.functional.classification.accuracy import binary_accuracy, multiclass_accuracy
 from torchmetrics.functional.classification.auroc import binary_auroc, multiclass_auroc
 from torchmetrics.functional.classification.confusion_matrix import (
@@ -71,6 +80,42 @@ from torchmetrics.utilities.plot import plot_binary_roc_curve, plot_confusion_ma
             lambda: torch.randint(3, (100,)),
             id="multiclass and average=None",
         ),
+        pytest.param(
+            partial(perceptual_evaluation_speech_quality, fs=8000, mode="nb"),
+            lambda: torch.randn(8000),
+            lambda: torch.randn(8000),
+            id="perceptual_evaluation_speech_quality",
+        ),
+        pytest.param(
+            partial(signal_distortion_ratio),
+            lambda: torch.randn(8000),
+            lambda: torch.randn(8000),
+            id="signal_distortion_ratio",
+        ),
+        pytest.param(
+            partial(scale_invariant_signal_distortion_ratio),
+            lambda: torch.randn(5),
+            lambda: torch.randn(5),
+            id="scale_invariant_signal_distortion_ratio",
+        ),
+        pytest.param(
+            partial(signal_noise_ratio),
+            lambda: torch.randn(4),
+            lambda: torch.randn(4),
+            id="signal_noise_ratio",
+        ),
+        pytest.param(
+            partial(scale_invariant_signal_noise_ratio),
+            lambda: torch.randn(4),
+            lambda: torch.randn(4),
+            id="scale_invariant_signal_noise_ratio",
+        ),
+        pytest.param(
+            partial(short_time_objective_intelligibility, fs=8000, extended=False),
+            lambda: torch.randn(8000),
+            lambda: torch.randn(8000),
+            id="short_time_objective_intelligibility",
+        ),
     ],
 )
 @pytest.mark.parametrize("num_vals", [1, 5, 10])
@@ -78,6 +123,28 @@ def test_single_multi_val_plotter(metric, preds, target, num_vals):
     vals = []
     for i in range(num_vals):
         vals.append(metric(preds(), target()))
+    vals = vals[0] if i == 1 else vals
+    fig, ax = plot_single_or_multi_val(vals)
+    assert isinstance(fig, plt.Figure)
+    assert isinstance(ax, matplotlib.axes.Axes)
+
+
+@pytest.mark.parametrize(
+    ("metric", "preds", "target"),
+    [
+        pytest.param(
+            partial(permutation_invariant_training, metric_func=scale_invariant_signal_noise_ratio, eval_func="max"),
+            lambda: torch.randn(3, 2, 5),
+            lambda: torch.randn(3, 2, 5),
+            id="permutation_invariant_training",
+        )
+    ],
+)
+@pytest.mark.parametrize("num_vals", [1, 5, 10])
+def test_single_multi_val_plotter_pit(metric, preds, target, num_vals):
+    vals = []
+    for i in range(num_vals):
+        vals.append(metric(preds(), target())[0])
     vals = vals[0] if i == 1 else vals
     fig, ax = plot_single_or_multi_val(vals)
     assert isinstance(fig, plt.Figure)